refactor command folder structure
