+++ js/commands/utilities/ping/ping.js
